Implement an OpenAPI interface for AuthorizationService (formerly RuleReaderService)
RuleReaderService was strictly confined to run inside the
GeoServer plugin.

A new "Authorization" API is introduced to the OpenAPI
specification, with client and server implementations,
and a client adaptor
AuthorizationServiceClientAdaptor implements AuthorizationService,
which calls the remote service to fetch the grant access info.

Now the Authorization service runs on the REST application
service and is exposed through the REST API, and the client
can be configured to wither run in-process as it always did
in GeoServer with GeoFence, or to call the remote API instead.

This allows to have the cache centralized on the server
and alleviate the GeoServer instances from that burden,
especially when there are several GeoServer instances
with the plugin running (target scenario is in the order of tenths).