Authorization api refactoring
* Move authorization out of domain to application. In terms
of architecture, it's purely use cases, whilst rule and
adminrule management is domain.

* Replace usage for RuleFilter by AccessRequest. RuleFilter
belongs to the rule administration domain. The use cases
that the authorization api handles require a clear request/response
interface that has nothing to do with RuleFilter's ANY, DEFAULT,
and all that machinery. The semantics of an AccessRequest are
just related to a given user performing a request, and the api
must return the AccessInfo grants, nothing else. Converting
an AccessRequest to a RuleFilter or a set of RuleFilters is
performed by the AuthorizationService implementation.